Prevent potential overflow when creating color intensities
Fixes #380

Recent changes appear to trigger overflow protection code that crashes on latest iOS + devices. Doesn't occur on debug builds, only release. The issue is that components for setting color intensities are declared as int8_t (signed) when they shouldn't be. See linked issue for more explanation, but the fix is to a) declare proper storage type and b) clamp component values to [0.0, 1.0].